requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
packaging==25.0
pefile==2023.2.7
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyshorteners==1.0.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
